LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/setrequestheader-content-type-expected.txt


FAIL XMLHttpRequest: setRequestHeader() - Content-Type header () assert_equals: expected "content-type: \n" but got "content-type: , text/plain;charset=UTF-8\n"
FAIL XMLHttpRequest: setRequestHeader() - Content-Type header ( ) assert_equals: expected "content-type: \n" but got "content-type: , text/plain;charset=UTF-8\n"
PASS XMLHttpRequest: setRequestHeader() - Content-Type header (null) 
PASS XMLHttpRequest: setRequestHeader() - Content-Type header (undefined) 
